Edit: FeedbackService_v1.1.0

+++ back-end/src/main/java/com/example/demo/dto/Feedback/ReplyRequestDto.java
@@ -8,8 +8,8 @@ public class ReplyRequestDto {
 
     private String reply;
 
-    private Integer statusId;
+    private Long statusId;
 
-    private Integer adminId;
+    private Long adminId;
 
 }
