+++ back-end/src/main/java/com/example/demo/repository/Feedback/FeedbackStatusRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.demo.entity.Feedback.FeedbackStatus;
 
-public interface FeedbackStatusRepository extends JpaRepository<FeedbackStatus, Integer> {
+public interface FeedbackStatusRepository extends JpaRepository<FeedbackStatus, Long> {
 
 }
